+++ src/oca/go/src/goca/marketplaceapp.go
@@ -35,7 +35,7 @@ type MarketPlaceAppPool struct {
 
 // MarketPlaceApp represents an OpenNebula MarketPlaceApp
 type MarketPlaceApp struct {
-	ID            uint              `xml:"ID"`
+	ID            int               `xml:"ID"`
 	UID           int               `xml:"UID"`
 	GID           int               `xml:"GID"`
 	UName         string            `xml:"UNAME"`
@@ -93,13 +93,13 @@ func (c *Controller) MarketPlaceApps() *MarketPlaceAppsController {
 }
 
 // MarketPlaceApp returns a MarketPlaceApp controller
-func (c *Controller) MarketPlaceApp(id uint) *MarketPlaceAppController {
+func (c *Controller) MarketPlaceApp(id int) *MarketPlaceAppController {
 	return &MarketPlaceAppController{c, id}
 }
 
 // ByName returns a MarketPlace ID from name
-func (c *MarketPlaceAppsController) ByName(name string, args ...int) (uint, error) {
-	var id uint
+func (c *MarketPlaceAppsController) ByName(name string, args ...int) (int, error) {
+	var id int
 
 	marketAppPool, err := c.Info(args...)
 	if err != nil {
@@ -179,7 +179,7 @@ func (mc *MarketPlaceAppController) Info() (*MarketPlaceApp, error) {
 // * name: name of the app
 // * tpl: template of the marketplace app
 // * market: market place ID
-func (mc *MarketPlaceAppsController) Create(name string, tpl *MarketAppTemplate, marketID int) (uint, error) {
+func (mc *MarketPlaceAppsController) Create(name string, tpl *MarketAppTemplate, marketID int) (int, error) {
 	if tpl == nil {
 		return 0, fmt.Errorf("MarketPlaceApp Create: nil template arg")
 	}
@@ -190,7 +190,7 @@ func (mc *MarketPlaceAppsController) Create(name string, tpl *MarketAppTemplate,
 		return 0, err
 	}
 
-	return uint(response.BodyInt()), nil
+	return response.BodyInt(), nil
 }
 
 // Delete deletes the given marketplace app from the pool.
